requirements.txt
numpy
openai>=1.0.0
python-dotenv

